Extracted post attr gating to content-gating service
no issue

- moved remaining gating logic to content-gating service
- kept modify-by-reference to avoid deep cloning post objects for performance on a hot path
- cleaned up requires to avoid circular dependencies and function wrappers
